ClientTransactionImpl should support timer C as well.
Support for timer C was implemented as Proxy/Branch timeout.

However for cases when we act as UAC that was not enough, as Client
Transactions were left hanging around for 12min (keeping Sip Sessions
and Sip Application Sessions alive together with transaction) for cases
when timer C would fire (e.g. we received provisional response for
INVITE, but no final response in 3 minutes).